requirements.txt
# Core packages
langchain==0.1.16
langchain-core==0.1.42          # ✅ Updated to match langchain 0.1.16
langchain-openai==0.0.2
langchain-pinecone==0.0.3

# Embeddings and LLMs
openai
pinecone-client

# PDF support
PyMuPDF

# Utilities
python-dotenv
tiktoken
pydantic==1.10.8
flask                         # ✅ Add this so it's available

# ML
tensorflow-cpu
scikit-learn
numpy
pandas

#DB
pymongo